refactor: 사이드바 아아ㅣ템 폴더 생성
